Result set

+++ JdbcPrimer/src/model/Kurs.java
@@ -4,7 +4,7 @@ public class Kurs {
 	
 	private int idKursa;
 	private String imeKursa;
-	private int cena;
+	private double cena;
 	
 	public int getIdKursa() {
 		return idKursa;
@@ -18,10 +18,10 @@ public class Kurs {
 	public void setImeKursa(String imeKursa) {
 		this.imeKursa = imeKursa;
 	}
-	public int getCena() {
+	public double getCena() {
 		return cena;
 	}
-	public void setCena(int cena) {
+	public void setCena(double cena) {
 		this.cena = cena;
 	}
